+++ packages/react-openapi/src/OpenAPIRequestBody.tsx
@@ -1,8 +1,10 @@
 import type { OpenAPIV3 } from '@gitbook/openapi-parser';
 import { InteractiveSection } from './InteractiveSection';
 import { OpenAPIRootSchema } from './OpenAPISchemaServer';
-import type { OpenAPIClientContext, OpenAPIOperationData } from './types';
-import { checkIsReference } from './utils';
+import type { OpenAPIClientContext } from './context';
+import { t } from './translate';
+import type { OpenAPIOperationData, OpenAPIWebhookData } from './types';
+import { checkIsReference, createStateKey } from './utils';
 
 /**
  * Display an interactive request body.
@@ -10,10 +12,9 @@ import { checkIsReference } from './utils';
 export function OpenAPIRequestBody(props: {
     requestBody: OpenAPIV3.RequestBodyObject | OpenAPIV3.ReferenceObject;
     context: OpenAPIClientContext;
-    data: OpenAPIOperationData;
+    data: OpenAPIOperationData | OpenAPIWebhookData;
 }) {
     const { requestBody, context, data } = props;
-    const { method, path } = data;
 
     if (checkIsReference(requestBody)) {
         return null;
@@ -21,9 +22,9 @@ export function OpenAPIRequestBody(props: {
 
     return (
         <InteractiveSection
-            header="Body"
+            header={t(context.translation, 'name' in data ? 'payload' : 'body')}
             className="openapi-requestbody"
-            stateKey={`media-type-${method}-${path}`}
+            stateKey={createStateKey('request-body-media-type', context.blockKey)}
             selectIcon={context.icons.chevronDown}
             tabs={Object.entries(requestBody.content ?? {}).map(
                 ([contentType, mediaTypeObject]) => {