+++ packages/react-openapi/src/ScalarApiButton.tsx
@@ -6,6 +6,8 @@ import { createPortal } from 'react-dom';
 
 import type { OpenAPIV3_1 } from '@gitbook/openapi-parser';
 import { useOpenAPIOperationContext } from './OpenAPIOperationContext';
+import type { OpenAPIClientContext } from './context';
+import { t } from './translate';
 
 /**
  * Button which launches the Scalar API Client
@@ -14,8 +16,9 @@ export function ScalarApiButton(props: {
     method: OpenAPIV3_1.HttpMethods;
     path: string;
     specUrl: string;
+    context: OpenAPIClientContext;
 }) {
-    const { method, path, specUrl } = props;
+    const { method, path, specUrl, context } = props;
     const [isOpen, setIsOpen] = useState(false);
     const controllerRef = useRef<ScalarModalControllerRef>(null);
     return (
@@ -27,7 +30,7 @@ export function ScalarApiButton(props: {
                     setIsOpen(true);
                 }}
             >
-                Test it
+                {t(context.translation, 'test_it')}
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 10 12" fill="currentColor">
                     <path
                         stroke="currentColor"